backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Environment & Configuration
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
python-magic==0.4.27
Pillow==10.1.0

# Document Processing
langchain==0.0.350
langchain-text-splitters==0.0.1
chromadb==0.4.22
underthesea==1.3.5

# Vector Store
qdrant-client==1.7.0

# Utilities
python-magic-bin==0.4.14; sys_platform == 'win32'

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 